+++ lib/screens/home_screens/home_screen.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:project_lagann/screens/home_screens/home_all_screen.dart';
-import 'package:project_lagann/screens/notifications_screen.dart';
+import 'package:project_lagann/screens/notification_screens/notifications_screen.dart';
 import 'package:project_lagann/screens/search_screen.dart';
 import 'package:project_lagann/utils/constants.dart';
 import '../../generated/l10n.dart';
+import '../../widgets/home_widgets/custom_home_appbar.dart';
 import 'home_followed_screen.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -132,26 +133,4 @@ class _HomeScreenState extends State<HomeScreen> {
     _pageController.animateToPage(1,
         duration: const Duration(milliseconds: 500), curve: Curves.decelerate);
   }
-}
-
-class CustomHomeAppBar extends StatelessWidget implements PreferredSizeWidget {
-  final double height;
-  final AppBar appBar;
-  final Widget otherWidget;
-  const CustomHomeAppBar(
-      {Key? key,
-      required this.appBar,
-      required this.otherWidget,
-      required this.height})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [appBar, otherWidget],
-    );
-  }
-
-  @override
-  Size get preferredSize => Size.fromHeight(height);
 }
